model "Review" extracted from apps.products to apps.reviews, POST and GET method created, control checked
