+++ netomerch/apps/products/management/commands/load_init_item.py
@@ -32,4 +32,3 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         self.handle1(args, options)
-
