+++ netomerch/apps/products/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 from taggit.serializers import TaggitSerializer, TagListSerializerField
 
-from apps.products.models import Category, Image, Item, Review
+from apps.products.models import Category, Image, Item
 
 
 class CategorySerializer(serializers.ModelSerializer):
@@ -28,23 +28,3 @@ class ItemSerializer(TaggitSerializer, serializers.ModelSerializer):
     class Meta:
         model = Item
         fields = ("id", "name", "short_description", "description", "image", "tags", "category", "properties")
-
-
-class ReviewSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Review
-        fields = ('id', 'item_id', 'author', 'email', 'text', 'is_published')
-
-
-class SendReviewSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Review
-        fields = ('id', 'item', 'author', 'email', 'text')
-
-    def validate_item(self, data):
-        print(f"\n\n\n\n{data}")
-        if data is None:
-            raise serializers.ValidationError('field item couldn''t be empty')
-        return data
