+++ netomerch/apps/products/urls.py
@@ -1,10 +1,9 @@
 from rest_framework.routers import DefaultRouter
 
-from apps.products.views import CategoryViewSet, ItemViewSet, ReviewViewSet
+from apps.products.views import CategoryViewSet, ItemViewSet
 
 router = DefaultRouter()
 router.register('categories', CategoryViewSet, basename='categories')
 router.register('items', ItemViewSet, basename='items')
-router.register('reviews', ReviewViewSet, basename='reviews')
 
 urlpatterns = router.urls
