+++ netomerch/apps/reviews/admin.py
@@ -1,0 +1,10 @@
+from django.contrib import admin
+
+from apps.reviews.models import Review
+
+# Register your models here.
+
+
+@admin.register(Review)
+class ReviewAdmin(admin.ModelAdmin):
+    pass
