+++ netomerch/config/urls.py
@@ -9,6 +9,7 @@ urlpatterns = api_docs_urlpatterns + [
     path("admin/", admin.site.urls),
     path("api/v1/", include("apps.products.urls")),
     path("api/v1/", include("apps.orders.urls")),
+    path("api/v1/", include("apps.reviews.urls")),
     path("api/v1/task/", include("apps.taskqueue.urls")),
     path("callback/", include("apps.email.urls")),
     path('summernote/', include('django_summernote.urls')),
